Foursquare for Make Your Day, with bugs
All that needs to be done, is filling in the prompt with the data we actually want to send to gpt (in a good format for it). The bug though, is that the geocoding will always take string vars but not user input for some reason. Will try to fix later.