Load more results button
user can load more results, and newer results appear first in the list. Unfortunately, the limit is not per activity, it is for all of them (id have to do some gutting to fix that I think). Technically if you click load more than like 10 times, it will stop giving you results at all. I can fix that by tracking it and adjusting lat/long when the limit is reached, but that's unnecessary for now.

Also, I noticed that the pins should probably have a label over them when you click a card.

And lastly, we are getting results from query rather than categories unless I'm mistaken. It might be working ok, but it definitely isn't ideal.